Fix bug with large files reporting wrong size
Large files would cause the output to show 4194303 kb transferred, or
the max value of an unsigned int / 1024.
Not sure why this happens, as the files weren't that big.
Changed size in main.c from unsigned int to unsigned long int.

+++ titledumper/source/main.c
@@ -195,11 +195,11 @@ void processTag(int client_socket, SendData *sendData)
         if(!pFile) {
             break;
         }
-        unsigned int size = ftell(pFile);
+        unsigned long int size = ftell(pFile);
         unsigned int time = gettime();
         float fTimeDiff = (time - lastTime) * 0.001f;
         float fSpeed = (fTimeDiff == 0.0f) ? 0.0f : ( (float)size / fTimeDiff / 1024.0f );
-        printf("Read file %i kb loaded with %0.3f kb/s\r", size / 1024, fSpeed);
+        printf("Read file %lu kb loaded with %0.3f kb/s\r", size / 1024, fSpeed);
         fwrite(sendData->data, 1, le32(sendData->length), pFile);
         break;
     }
